backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai>=1.30.0,<2
sentence-transformers
pypdf
tiktoken
numpy
scikit-learn
pdfminer.six
